+++ sigils/threadDefault.nim
@@ -35,6 +35,7 @@ method send*(
       raise newException(MessageQueueFullError, "could not send!")
     debugQueuePrint "queue:thread inputs size: ", $thread.inputs.peek(),
       " thread: ", $getThreadId(thread.toSigilThread()[])
+  thread.toSigilThread().notifyMessageEnqueued()
 
 method recv*(
     thread: SigilThreadDefaultPtr, msg: var ThreadSignal,