+++ sigils/threads.nim
@@ -17,6 +17,8 @@ import threadAsyncs
 when defined(feature.sigils.chronos):
   import threadChronos
 
+import threadExtras
+
 export isolateutils
 export threadBase
 export threadDefault
@@ -24,4 +26,5 @@ export threadPool
 export threadProxies
 when defined(feature.sigils.chronos):
   export threadChronos
+export threadExtras
 # export threadAsyncs